Increase max_length of Code.code to 30
This commit includes the corresponding schema migration.

The police API uses hyphenated area names as force IDs, many of
which are longer than 10 characters (the previous value of
max_length). The longest is currently 18 characters, so 30 should
be plenty for this and hopefully also for any other new code types.
Performance should be unaffected by this change:

  http://www.postgresql.org/docs/9.1/static/datatype-character.html

I decided not to remove the length constraint by switching to a
TextField because that could affect how Django displays it in the
admin interface.