Remove add_new_two_d_polygon
Trying to create a Geometry with a polygon fails because the
polygon is 3D and the Polygon field has dim=2 by default; use
polygon.wkt instead, which appears to be 2D now.

This also removes my excellent third-co-ordinate-matching
regular expression, sadly.

Commit 0a5cd8b2802f68e259eb9c4d75b7341ec18afa10 removed
must_be_two_d from save_polygons in utils.py, which also
suggests that this is no longer necessary, although no
explanation is given there for that change.