Add police copyright and licence info to GB copyright template
Both the API data and neighbourhood KMLs are licensed under the
Open Government Licence. At the bottom of the page from which the
neighbourhood KMLs are available along with many other datasets:

    http://www.police.uk/data

...it states 'Contains Ordnance Survey data (c) Crown copyright and
database right 2011.' It's not clear if this is applies to all
datasets available there, but I've added 'and 2011' to the existing
line about Ordnance Survey data anyway.