Add police force areas for England and Wales
Police areas in England and Wales are defined in Schedule 1
to the Police Act 1996 by descriptions of the administrative
areas which they cover, except for the two London forces,
which are defined elsewhere.

This commit includes a data file which gives for each force:
  - the code used by the Police API as an ID
  - its full name, available via the Police API
  - the description of its area from the Police Act
  - a set of lookups which uniquely identifies each covered
    administrative area in the Boundary-Line data using its
    type code and name

The script included here uses this data to create areas for
police forces and to associate them with the appropriate
polygons.