Increase max_length of code, name and description model fields
The maximum lengths set for many fields are shorter than the
data we might want to store in them, for example:

  - ISO country codes are 2 or 3 characters long, but
    Country.code has max_length=1
  - Various countries including Argentina, Brazil, Iran,
    Lebanon and the US use postcode formats which are
    longer than 7 characters, the current max_length of
    Postcode.postcode
  - Force codes used by the Police API for police forces in
    England and Wales are up to 18 characters long, but
    Code.code has max_length=10

This commit increases the max_length of all code fields
(including postcode) to 100, and all name and description
fields to 255 characters. 255 is the maximum supported by
MySQL for VARCHAR columns with unique=True:

  https://docs.djangoproject.com/en/1.2/ref/databases/#character-fields

Performance should be unaffected by this change:

  http://www.postgresql.org/docs/9.1/static/datatype-character.html

This commit also includes a schema migration for these changes, and
a corresponding empty data migration.

The reason for the empty data migration is that before reversing
migration 0008, any codes, names or descriptions which are longer
than the earlier value of max_length for their fields will need to
be shortened. Simply truncating values might attempt to create
duplicate values in fields with unique=True, so developers will
need to write their own backwards methods in this empty placeholder
data migration to shorten them appropriately.